Port to pytest, add SHAP library, cleanup
I replaced `unittest` with `pytest` for tests. They can be run by
first installing the repository as an editable python package via
`python -m pip install -e .`. Then simply running `pytest` in the
root directory of the repo runs the tests.

I completely ported `test_fd_imputer.py`, however I didn't port
`test_dep_detector.py` yet. The tests are mostly integration tests,
and a few unit tests.

I also changed the behaviour of `split_dataset`: I previously stored the
ordering of the rows whenever I splitted the datasets. This was
initially thought of as a sanity test -- the fd "ID -> row" should
always exist.

However, in the logic of PFDs, this adds a new feature to the dataset
that really shouldn't be there. Which is why I removed this. The
binaries in the `MLFD_fd_detection/` folder still need to be adjusted to
this change.
