+++ plotter.py
@@ -5,7 +5,6 @@ import matplotlib.pyplot as plt
 from matplotlib.ticker import MaxNLocator
 import numpy as np
 import pickle
-import lib.fd_imputer as fd
 import lib.plot_utils as pu
 import lib.constants as c
